97453 - updates to card
