+++ src/applications/accredited-representative-portal/sass/POARequestsCard.scss
@@ -1,0 +1,43 @@
+@import "~@department-of-veterans-affairs/css-library/dist/tokens/scss/variables";
+
+.poa-request {
+  &__list {
+    list-style-type: none;
+    padding: 0;
+    display: block;
+  }
+
+  &__card {
+    margin-bottom: 20px;
+
+    @media (min-width: $small-desktop-screen) {
+      max-width: 600px;
+    }
+
+    &-field,
+    &-title {
+      margin: 0;
+
+      &--status {
+        background: $vads-color-base-dark;
+        color: var(--vads-color-white);
+        max-width: fit-content;
+        text-transform: uppercase;
+        padding: 2px 10px;
+        margin-bottom: 5px;
+      }
+
+      &--request {
+        display: flex;
+        align-items: center;
+        margin-top: 8px;
+      }
+
+      &--label {
+        font-weight: 700;
+        margin-bottom: 0;
+        margin-right: 5px;
+      }
+    }
+  }
+}
